fix: extra functionality not called
